Moved Rat into algebra.laws test scope
Also moved Rat arbitrary directly into Rat companion object. CheckSupport
is now empty. Remove?
