+++ laws/jvm/src/test/scala/algebra/laws/LawTestsJvm.scala
@@ -2,8 +2,6 @@ package algebra
 package laws
 
 import algebra.macros._
-import algebra.std.Rat
-import CheckSupport._
 
 class LawTests extends LawTestsBase {
